server/boardServer/requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
kafka-python==2.0.2
Pillow==10.0.0
pytz==2023.3
redis==4.6.0
sqlparse==0.4.4
tzdata==2023.3
